@@ -1,5 +1,7 @@
-describe('Dummy', () => {
-  test('should work', () => {
+describe("Dummy", () => {
+  test("should work", () => {
     expect(42).toBe(42);
   });
 });
+
+export const dummy = {};
